[ISSUE #4236]♻️Refactor error handling to use RocketMQError consistently; improve clarity and maintainability across multiple files

+++ rocketmq-client/src/implementation/mq_client_api_impl.rs
@@ -2192,7 +2192,7 @@ impl MQClientAPIImpl {
                 }
                 code => {
                     return Err(mq_client_err!(
-                        response.code() as i32,
+                        response.code(),
                         response.remark().map_or("".to_string(), |s| s.to_string())
                     ));
                 }
